+++ src/TwigComponent/src/Twig/ComponentNode.php
@@ -15,7 +15,9 @@ use Symfony\UX\TwigComponent\BlockStack;
 use Twig\Attribute\YieldReady;
 use Twig\Compiler;
 use Twig\Environment;
+use Twig\Error\SyntaxError;
 use Twig\Node\Expression\AbstractExpression;
+use Twig\Node\Expression\NameExpression;
 use Twig\Node\Node;
 use Twig\Node\NodeOutputInterface;
 use Twig\Template;
@@ -29,9 +31,9 @@ use Twig\Template;
 #[YieldReady]
 final class ComponentNode extends Node implements NodeOutputInterface
 {
-    public function __construct(string $component, string $embeddedTemplateName, int $embeddedTemplateIndex, ?AbstractExpression $props, bool $only, int $lineno)
+    public function __construct(AbstractExpression $component, string $embeddedTemplateName, int $embeddedTemplateIndex, ?AbstractExpression $props, bool $only, int $lineno)
     {
-        $nodes = [];
+        $nodes = ['component' => $component];
         if (null !== $props) {
             $nodes['props'] = $props;
         }
@@ -41,7 +43,6 @@ final class ComponentNode extends Node implements NodeOutputInterface
         $this->setAttribute('only', $only);
         $this->setAttribute('embedded_template', $embeddedTemplateName);
         $this->setAttribute('embedded_index', $embeddedTemplateIndex);
-        $this->setAttribute('component', $component);
     }
 
     public function compile(Compiler $compiler): void
@@ -56,6 +57,44 @@ final class ComponentNode extends Node implements NodeOutputInterface
                ->string(ComponentRuntime::class)
                ->raw(");\n");
 
+        $componentName = $compiler->getVarName();
+        $componentExpression = $this->getNode('component');
+
+        if ($componentExpression instanceof NameExpression && !$componentExpression->hasExplicitParentheses()) {
+            $compiler
+                ->write(\sprintf('$%s = %s;', $componentName, var_export($componentExpression->getAttribute('name'), true)))
+                ->raw("\n");
+        } else {
+            $componentNameValue = $compiler->getVarName();
+            $compiler
+                ->write(\sprintf('$%s = ', $componentNameValue))
+            ;
+
+            $compiler->subcompile($componentExpression);
+            $compiler->raw(";\n");
+
+            $compiler
+                ->write(\sprintf('if (\\is_scalar($%s) || $%s instanceof \\Stringable) {', $componentNameValue, $componentNameValue))
+                ->raw("\n")
+                ->indent()
+                ->write(\sprintf('$%s = (string) $%s;', $componentName, $componentNameValue))
+                ->raw("\n")
+                ->outdent()
+                ->write("} else {\n")
+                ->indent()
+                ->write('throw new ')
+                ->raw('\\'.SyntaxError::class)
+                ->raw('(sprintf(')
+                ->string('The component expression passed to "{%% component %%}" must evaluate to a component name (string/scalar/Stringable). Got "%s".')
+                ->raw(', \\get_debug_type(')
+                ->raw(\sprintf('$%s', $componentNameValue))
+                ->raw(')), ')
+                ->repr($this->getTemplateLine())
+                ->raw(", \$this->getSourceContext());\n")
+                ->outdent()
+                ->write("}\n");
+        }
+
         /*
          * Block 1) PreCreateForRender handling
          *
@@ -64,7 +103,7 @@ final class ComponentNode extends Node implements NodeOutputInterface
          */
         $compiler
             ->write(\sprintf('$preRendered = $%s->preRender(', $componentRuntime))
-            ->string($this->getAttribute('component'))
+            ->raw(\sprintf('$%s', $componentName))
             ->raw(', ')
             ->raw('Twig\Extension\CoreExtension::toArray')
             ->raw('(');
@@ -96,7 +135,7 @@ final class ComponentNode extends Node implements NodeOutputInterface
          */
         $compiler
             ->write(\sprintf('$preRenderEvent = $%s->startEmbedComponent(', $componentRuntime))
-            ->string($this->getAttribute('component'))
+            ->raw(\sprintf('$%s', $componentName))
             ->raw(', ')
             ->raw('Twig\Extension\CoreExtension::toArray')
             ->raw('(');