+++ src/TwigComponent/tests/Fixtures/Component/DynamicNameComponent2.php
@@ -1,0 +1,19 @@
+<?php
+
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Symfony\UX\TwigComponent\Tests\Fixtures\Component;
+
+use Symfony\UX\TwigComponent\Attribute\AsTwigComponent;
+
+#[AsTwigComponent('DynamicNameComponent2', template: 'components/dynamic_name_component_2.html.twig')]
+final class DynamicNameComponent2
+{
+}